# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMH License Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface License {
  id: number;
  licenseKey: string;
  machineId: string;
  /** recorder | cash_sheeter | inventory | suite */
  product: string;
  /** trial | subscription | perpetual */
  licenseType: string;
  isActive: boolean;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface LicenseInput {
  /** @minLength 1 */
  machineId: string;
  /** recorder | cash_sheeter | inventory | suite */
  product: string;
  /** trial | subscription | perpetual */
  licenseType: string;
  /**
     * Override duration in days (default: 7 for trial, 30 for subscription, null for perpetual)
     * @nullable
     */
  durationDays?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface LicenseUpdate {
  isActive?: boolean;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface LicenseValidationInput {
  licenseKey: string;
  machineId: string;
  product: string;
}

export interface LicenseValidationResult {
  valid: boolean;
  message: string;
  /** @nullable */
  licenseType?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  daysRemaining?: number | null;
}

export interface ProductCount {
  product: string;
  count: number;
}

export interface LicenseStats {
  total: number;
  active: number;
  revoked: number;
  expired: number;
  trial: number;
  subscription: number;
  perpetual: number;
  byProduct: ProductCount[];
  recentLicenses: License[];
}

export type ListLicensesParams = {
product?: string;
licenseType?: string;
status?: string;
search?: string;
};