# lib/api-zod/src/generated/types/license.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMH License Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface License {
  id: number;
  licenseKey: string;
  machineId: string;
  /** recorder | cash_sheeter | inventory | suite */
  product: string;
  /** trial | subscription | perpetual */
  licenseType: string;
  isActive: boolean;
  /** @nullable */
  expiresAt?: Date | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}