# lib/api-zod/src/generated/types/licenseInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMH License Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface LicenseInput {
  /** @minLength 1 */
  machineId: string;
  /** recorder | cash_sheeter | inventory | suite */
  product: string;
  /** trial | subscription | perpetual */
  licenseType: string;
  /**
     * Override duration in days (default: 7 for trial, 30 for subscription, null for perpetual)
     * @nullable
     */
  durationDays?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  notes?: string | null;
}