# lib/api-zod/src/generated/types/licenseStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMH License Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { License } from './license';
import type { ProductCount } from './productCount';

export interface LicenseStats {
  total: number;
  active: number;
  revoked: number;
  expired: number;
  trial: number;
  subscription: number;
  perpetual: number;
  byProduct: ProductCount[];
  recentLicenses: License[];
}