# lib/api-zod/src/generated/types/licenseUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMH License Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface LicenseUpdate {
  isActive?: boolean;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  notes?: string | null;
}