# lib/api-zod/src/generated/types/licenseValidationResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMH License Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface LicenseValidationResult {
  valid: boolean;
  message: string;
  /** @nullable */
  licenseType?: string | null;
  /** @nullable */
  expiresAt?: Date | null;
  /** @nullable */
  daysRemaining?: number | null;
}