# lib/api-zod/src/generated/types/productCount.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMH License Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductCount {
  product: string;
  count: number;
}